+++ packages/medical-domain/src/domains/repositories/medical/datum/zen-mode.service.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2022-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -44,7 +44,8 @@ const normalize = (rawData: Record<string, unknown>, opts: MedicalDataOptions): 
     type: DatumType.DeviceEvent,
     subType: DeviceEventSubtype.Zen,
     guid: rawData.guid as string,
-    inputTime: rawData.inputTime as string
+    inputTime: rawData.inputTime as string,
+    glycemiaTarget: null
   }
   return zenMode
 }